feat(checkbox): replace legacy checkbox with alpha checkbox
