--- packages/bezier-react/src/components/Forms/AlphaCheckbox/index.ts
@@ -1,6 +1,0 @@
-export { AlphaCheckbox } from './AlphaCheckbox'
-
-export type {
-  CheckedState,
-  CheckboxProps as AlphaCheckboxProps,
-} from './AlphaCheckbox.types'
