--- packages/bezier-react/src/components/Forms/Checkbox/CheckType.ts
@@ -1,7 +1,0 @@
-enum CheckType {
-  False = 'false',
-  True = 'true',
-  Partial = 'partial',
-}
-
-export default CheckType
