+++ packages/bezier-react/src/components/Forms/Checkbox/Checkbox.types.ts
@@ -1,22 +1,44 @@
 /* Internal dependencies */
-import {
-  BezierComponentProps,
-  ChildrenProps,
-  AdditionalStylableProps,
-  AdditionalTestIdProps,
-} from 'Types/ComponentProps'
-import type { FormComponentProps } from 'Components/Forms/Form.types'
-import CheckType from './CheckType'
+import { BezierComponentProps, ChildrenProps } from 'Types/ComponentProps'
+import { FormComponentProps } from 'Components/Forms'
+
+export type CheckedState = boolean | 'indeterminate'
 
 interface CheckboxOptions {
-  checked?: boolean | CheckType
-  onClick?: React.MouseEventHandler
+  /**
+   * The controlled checked state of the checkbox.
+   * Must be used in conjunction with `onCheckedChange`.
+   */
+  checked?: CheckedState
+  /**
+   * The checked state of the checkbox when it is initially rendered.
+   * Use when you do not need to control its checked state.
+   */
+  defaultChecked?: CheckedState
+  /**
+   * The unique id of the checkbox. It is created automatically by default.
+   * It used by the label element in the checkbox.
+   */
+  id?: string
+  /**
+   * The name of the checkbox.
+   * Submitted with its owning form as part of a name/value pair.
+   */
+  name?: string
+  /**
+   * The value given as data when submitted with a name.
+   * @default 'on'
+   */
+  value?: string
+  /**
+   * Event handler called when the checked state of the checkbox changes.
+   */
+  onCheckedChange?: (checked: CheckedState) => void
 }
 
-export default interface CheckboxProps extends
+export interface CheckboxProps extends
   BezierComponentProps,
   ChildrenProps,
   FormComponentProps,
-  AdditionalStylableProps<'content'>,
-  AdditionalTestIdProps<'checker'>,
+  Omit<React.HTMLAttributes<HTMLButtonElement>, keyof CheckboxOptions>,
   CheckboxOptions {}
