+++ packages/bezier-react/src/index.ts
@@ -15,7 +15,6 @@ export * from 'Components/Divider'
 export * from 'Components/Emoji'
 export * from 'Components/Forms'
 export * from 'Components/Forms/Checkbox'
-export * from 'Components/Forms/AlphaCheckbox'
 export * from 'Components/Forms/FormControl'
 export * from 'Components/Forms/FormGroup'
 export * from 'Components/Forms/FormHelperText'
